[고동현] 파일 추가 API 연동

+++ front/src/components/left_sidebar/Buttons.tsx
@@ -1,44 +1,42 @@
-import {json} from "stream/consumers";
-import React from "react";
+import React, {useState} from "react";
 import {AiOutlineDelete, AiOutlineFileAdd, AiOutlineFolderAdd} from "react-icons/ai";
-import {useRecoilState, useRecoilValue} from "recoil";
-import {recentSelected} from "../../recoil/sidebarState";
-import {filesState} from "../../recoil/imageState";
+import {useRecoilState, useRecoilValue, useSetRecoilState} from "recoil";
+import {recentSelected, recentSelectedDirectoryAtom} from "../../recoil/sidebarState";
+import {filesAtom, newFileParentIDAtom} from "../../recoil/fileState";
+import {FileType} from "../../common/types";
+import {getFileNodeById} from "../../common/commonFunctions";
 
 export default function Buttons() {
+	const [isToolTipOpened, setIsToolTipOpened] = useState(false);
 	const recentSelectedFileId = useRecoilValue(recentSelected);
-	const [_, setFiles] = useRecoilState(filesState);
+	const recentSelectedDirId = useRecoilValue(recentSelectedDirectoryAtom);
+	const [files, setFiles] = useRecoilState(filesAtom);
+	const setNewFileParentIdAtom = useSetRecoilState(newFileParentIDAtom);
 
-	const onClickAddFile = async (event:React.MouseEvent, isFile:boolean) => {
-		const body = {
-			description: "string",
-			fileType: "document",
-			name: "test",
-			parentId: recentSelectedFileId,
-		};
+	const onClickAddFile = async (event:React.MouseEvent, fileType: FileType) => {
+		event.stopPropagation();
+		setNewFileParentIdAtom({
+			fileType,
+			padding: 0,
+			parentID: recentSelectedDirId,
+		});
+		setIsToolTipOpened(false);
+	};
 
-		const fetchPromise = await fetch("http://localhost:8080/userDirectory",
-			// https://stackoverflow.com/questions/32500073/request-header-field-access-control-allow-headers-is-not-allowed-by-itself-in-pr
-			{
-				// mode: "cors",
-				// credentials: "same-origin",
-				method: "POST",
-				headers: {
-					"Access-Control-Allow-Origin": "*",
-					"accept": "*/*",
-					"Content-Type": "application/json",
-				},
-				body: JSON.stringify(body),
-			});
-
-		console.log(fetchPromise);
+	const toggleTooltip = function() {
+		setIsToolTipOpened(prevState => !prevState);
 	};
 
 	const onClickDeleteFile = async () => {
+		setIsToolTipOpened(false);
+		if (recentSelectedFileId === "" || parseInt(recentSelectedFileId, 10) === 0) {
+			return;
+		}
 		const fetchPromise = await fetch(`http://localhost:8080/userDirectory/${recentSelectedFileId}`,
 			{
 				method: "DELETE",
-			});
+			},
+		);
 
 		if (fetchPromise.ok && fetchPromise.status === 200) {
 			setFiles(prev => prev.filter(value => value.getID !== recentSelectedFileId));
@@ -49,15 +47,27 @@ export default function Buttons() {
 	};
 
 	return <div id="left-sidebar-buttons">
-		<span onClick={e => onClickAddFile(e, false)} role={"button"}>
+		<span onClick={e => onClickAddFile(e, FileType.directory)} role={"button"}>
 			<AiOutlineFolderAdd/>
 		</span>
-		<span onClick={e => onClickAddFile(e, true)} role={"button"}>
+
+		<span id="new-file" role={"button"} onClick={toggleTooltip}>
 			<AiOutlineFileAdd/>
+			<div id={"new-file-tooltip"}
+				style={{
+					display: isToolTipOpened ? "grid" : "none",
+				}}
+			>
+				<span onClick={e => onClickAddFile(e, FileType.document)}>
+					Document
+				</span>
+				<span onClick={e => onClickAddFile(e, FileType.task_file)}>
+					Task
+				</span>
+			</div>
 		</span>
 		<span onClick={onClickDeleteFile} role={"button"}>
 			<AiOutlineDelete/>
 		</span>
-
 	</div>;
 }
